requirements.txt
# --- Frameworks Principales ---
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.9 # Requerido por FastAPI para los formularios de login

# --- Base de Datos y ORM (Fase 3) ---
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-decouple==3.8

# --- Seguridad (Fase 3) ---
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# --- Lógica del Núcleo de IA ---
experta==1.9.4
rank-bm25==0.2.2
scikit-learn==1.7.1
spacy==3.8.7
networkx==3.5
numpy  # Permitir que pip resuelva la versión compatible

# --- Utilidades Generales ---
requests==2.32.5
python-dotenv==1.1.1

# --- Testing ---
pytest==8.4.1
psutil
cryptography
flake8
